Enum shared_library::LoadingError

pub enum LoadingError {
    LibraryNotFound {
        descr: String,
    },
    SymbolNotFound {
        symbol: &'static str,
    },
}

Error that can happen while loading the shared library.

Variants
LibraryNotFound

Fields of LibraryNotFound

descr: String

SymbolNotFound

One of the symbols could not be found in the library.

Fields of SymbolNotFound

symbol: &'static str

The symbol.