Struct shared_library::dynamic_library::DynamicLibrary

pub struct DynamicLibrary {
    // some fields omitted
}

Methods

impl DynamicLibrary

fn open(filename: Option<&Path>) -> Result<DynamicLibrary, String>

Lazily open a dynamic library. When passed None it gives a handle to the calling process

fn prepend_search_path(path: &Path)

Prepends a path to this process's search path for dynamic libraries

fn create_path(path: &[PathBuf]) -> OsString

From a slice of paths, create a new vector which is suitable to be an environment variable for this platforms dylib search path.

fn envvar() -> &'static str

Returns the environment variable for this process's dynamic library search path

fn search_path() -> Vec<PathBuf>

Returns the current search path for dynamic libraries being used by this process

unsafe fn symbol<T>(&self, symbol: &str) -> Result<*mut T, String>

Access the value at the symbol of the dynamic library