Struct shared_library::dynamic_library::DynamicLibrary

pub struct DynamicLibrary { /* private fields */ }

Implementations

impl DynamicLibrary

pub fn open(filename: Option<&Path>) -> Result<Self, String>

Lazily loads the dynamic library named filename into memory and then returns an opaque “handle” for that dynamic library.

Returns a handle to the calling process when passed None.

pub fn prepend_search_path(path: &Path)

Prepends a path to this process’s search path for dynamic libraries

pub fn create_path(path: &[PathBuf]) -> OsString

From a slice of paths, create a new vector which is suitable to be an environment variable for this platforms dylib search path.

pub fn envvar() -> &'static str

Returns the environment variable for this process’s dynamic library search path

pub fn search_path() -> Vec<PathBuf>

Returns the current search path for dynamic libraries being used by this process

pub unsafe fn symbol<T>(&self, symbol: &str) -> Result<*mut T, String>

Returns the address of where symbol symbol was loaded into memory.

In POSIX compliant systems, we return ‘Err’ if the symbol was not found, in this library or any of the libraries that were automatically loaded when this library was loaded.

pub unsafe fn symbol_special<T>( handle: SpecialHandles, symbol: &str ) -> Result<*mut T, String>

Returns the address of the first occurance of symbol symbol using the default library search order if you use SpecialHandles::Default.

Returns the address of the next occurance of symbol symbol after the current library in the default library search order if you use SpecialHandles::Next.